requirements.txt
aiohttp>=3.9.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
rich>=13.7.0
google-generativeai>=0.3.0
jinja2>=3.1.0
webdriver-manager>=4.0.0
